src/requirements.txt
beautifulsoup4==4.12.3
celery==5.3.6
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django_cors_headers==4.3.0
django-celery-beat==2.5.0
django_jazzmin==2.6.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
python-dotenv==1.0.1
isort==5.12.0
lxml==5.1.0
pillow==10.0.1
psycopg2-binary==2.9.5
redis==5.0.1
requests==2.31.0